tests: reconcile cloud_retention scale test
tests/cloud_retention_test was a copy of its scale_test/ variant, scaled
down to be runnable in CI. This commit removes the original scale_test/
and parameterizes instead to avoid leaving around significantly
duplicated code.

This also marks the test debug-only, since the original was only run
with release binaries.
