+++ tests/rptest/tests/cloud_retention_test.py
@@ -8,7 +8,7 @@
 # by the Apache License, Version 2.0
 
 from ducktape.utils.util import wait_until
-from ducktape.mark import ok_to_fail
+from ducktape.mark import matrix
 from rptest.services.cluster import cluster
 from rptest.tests.prealloc_nodes import PreallocNodesTest
 from rptest.clients.types import TopicSpec
@@ -16,47 +16,75 @@ from rptest.clients.rpk import RpkTool
 from rptest.services.redpanda import SISettings, MetricsEndpoint
 from rptest.services.kgo_verifier_services import (
     KgoVerifierConsumerGroupConsumer, KgoVerifierProducer)
+from rptest.utils.mode_checks import skip_debug_mode
 
 
 class CloudRetentionTest(PreallocNodesTest):
-    segment_size = 1024 * 1024
     default_retention_segments = 2
     topic_name = "si_test_topic"
 
     def __init__(self, test_context):
+        super(CloudRetentionTest, self).__init__(test_context=test_context,
+                                                 node_prealloc_count=1,
+                                                 num_brokers=3)
+
+    def setUp(self):
+        # Defer cluster startup so we can tweak configs based on the type of
+        # test run.
+        pass
+
+    @cluster(num_nodes=4)
+    @matrix(max_consume_rate_mb=[20, None])
+    @skip_debug_mode
+    def test_cloud_retention(self, max_consume_rate_mb):
+        """
+        Test cloud retention with an ongoing workload. The consume load comes
+        in two flavours:
+        * an ordinary fast consumer that repeatedly consumes the tail few
+          segments
+        * a slow consumer that can't keep up with retention and has to start
+          over when retention overcomes it
+        """
+        if self.redpanda.dedicated_nodes:
+            num_partitions = 100
+            msg_size = 128 * 1024
+            msg_count = int(100 * 1024 * 1024 * 1024 / msg_size)
+            segment_size = 10 * 1024 * 1024
+        else:
+            num_partitions = 1
+            msg_size = 1024
+            msg_count = int(1024 * 1024 * 1024 / msg_size)
+            segment_size = 1024 * 1024
+
+        if max_consume_rate_mb is None:
+            max_read_msgs = 25000
+        else:
+            max_read_msgs = 2000
+
+        si_settings = SISettings(self.test_context,
+                                 log_segment_size=segment_size)
+        self.redpanda.set_si_settings(si_settings)
+
         extra_rp_conf = dict(retention_local_target_bytes_default=self.
-                             default_retention_segments * self.segment_size,
+                             default_retention_segments * segment_size,
                              log_segment_size_jitter_percent=5,
                              group_initial_rebalance_delay=300,
                              cloud_storage_segment_max_upload_interval_sec=60,
                              cloud_storage_housekeeping_interval_ms=10_000)
+        self.redpanda.add_extra_rp_conf(extra_rp_conf)
 
-        super(CloudRetentionTest, self).__init__(
-            test_context=test_context,
-            node_prealloc_count=1,
-            num_brokers=3,
-            si_settings=SISettings(test_context,
-                                   log_segment_size=self.segment_size),
-            extra_rp_conf=extra_rp_conf)
+        self.redpanda.start()
+        self._create_initial_topics()
 
-    @ok_to_fail  # https://github.com/redpanda-data/redpanda/issues/8251
-    @cluster(num_nodes=4)
-    def test_cloud_retention(self):
-        """
-        Test cloud retention with an ongoing workload.
-        """
-        num_partitions = 1
         rpk = RpkTool(self.redpanda)
         rpk.create_topic(topic=self.topic_name,
                          partitions=num_partitions,
                          replicas=3,
                          config={
                              "cleanup.policy": TopicSpec.CLEANUP_DELETE,
-                             "retention.bytes": 5 * self.segment_size,
+                             "retention.bytes": 5 * segment_size,
                          })
 
-        msg_size = 1024
-        msg_count = int(1024 * 1024 * 1024 / msg_size)
         producer = KgoVerifierProducer(self.test_context,
                                        self.redpanda,
                                        self.topic_name,
@@ -64,7 +92,6 @@ class CloudRetentionTest(PreallocNodesTest):
                                        msg_count=msg_count,
                                        custom_node=self.preallocated_nodes)
 
-        max_read_msgs = 25000
         consumer = KgoVerifierConsumerGroupConsumer(
             self.test_context,
             self.redpanda,
@@ -107,12 +134,12 @@ class CloudRetentionTest(PreallocNodesTest):
             try:
                 size = sum(obj.content_length
                            for obj in self.cloud_storage_client.list_objects(
-                               self.si_settings.cloud_storage_bucket))
+                               si_settings.cloud_storage_bucket))
                 self.logger.info(f"bucket size: {size}")
                 # check that for each partition there is more than 1
                 # and less than 10 segments in the cloud (generous limits)
-                return size >= self.segment_size * num_partitions \
-                    and size <= 10 * self.segment_size * num_partitions
+                return size >= segment_size * num_partitions \
+                    and size <= 10 * segment_size * num_partitions
             except Exception as e:
                 self.logger.warn(f"error getting bucket size: {e}")
                 return False
@@ -122,4 +149,4 @@ class CloudRetentionTest(PreallocNodesTest):
         consumer.wait()
         self.logger.info("finished consuming")
         assert consumer.consumer_status.validator.valid_reads > \
-            self.segment_size * num_partitions / msg_size
+            segment_size * num_partitions / msg_size
